Create ReadMe.md
For my final project, I want to create an audio visualizer that reacts to a lofi song I am creating as a different final project for Sound Production class. I want this audio visualizer to show the wave forms and how they react to the baseline and other elements that I use based on volume level. 

Other works that have served as inspiration include the audio visualizer that was listed in the Final Project Examples tab, as well as wanting to experience what it would be like to see a song, such as one who could experience synesthesia would (seeing certain colors when listening to music). As a visual artist, and someone who loves music, I've always been interested in audio visualizers and being able to see what I'm hearing in a creative, interactive way. This project would be a perfect opportunity to learn how to do that.

I know that I will have to learn how to make the waveform of this song appear and what all goes into making it interactive, as well as being able to form it into any shape I want. I will need to create and use functions to determine when certain elements of this visualizer react/change depending on the decibel levels in the song. I will also need to make the link interactive, where the viewer can play and pause the song whenever they like. I will also be learning how to include particles and make those interactive with the music as well, that way the visualizer will be more appealing to look at instead of it being a simple waveform. 

As for a work plan, I have already begun this project because I was so excited to start it, so I will say what my schedule has been so far. I have worked on this project for at least 1hr - 1hr 30min every other day in a week, as I am also doing other projects for other courses. This way, my workload for this class is balanced out with the rest and I am not spending too much time on it for one day. My project is close to done as of when I post this markdown file, I am just working on finishing up the lofi song and will replace the song I am currently using as a placeholder whenever it is done.

I have completed basically everything that I have needed for this project; every component of this audio visualizer such as the shape I want the waveforms to be in, the song that I want to play for this visualizer and being able to connect the two. I've also been able to add in particles and make them interactive with the placeholder song, and I have selected a nice background as well to give this project more atmosphere that would go along nicely with the song.

All that I have left to do is finish my new lofi song, replace the placeholder song, and then figure out a way to make the github link work, because so far it just has a "loading" screen that never ends. The link for the preview page that I have on Visual Studio Code works just fine, but I messaged  and sent my project to Cory about how I could fix the original link. I am just waiting on a response. If we can't figure it out I plan on sharing the link that does work instead of the Github link (if this is allowed).